+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.8.1]
+
+### Added
+
+- Added `Y8S3` to the `Season` enum.
+- Added new operator `Ram` to `Operater` enum.
+- Added game mode `Standard` to `GameMode` enum.
+
 ## [0.8.0]
 
 ### Changed